% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scKnockoff-package.R
\docType{package}
\name{scKnockoff-package}
\alias{scKnockoff-package}
\alias{scKnockoff}
\title{scKnockoff: Knockoff-based differential analysis for scRNA-seq data}
\description{
Provides methods for differential analysis of single-cell RNA sequencing
(scRNA-seq) data under the knockoff framework. The package implements
large-scale conditional independence screening to control false discovery
rates in high-dimensional settings, with support for imputing missing data
via sc-softImpute incorporating additional latent variables and a matrix
decomposition-based knockoff construction.
}
\details{
The scKnockoff package is designed for high-dimensional single-cell data,
where the number of features can be much larger than the number of samples.
It supports missing data imputation using sc-softImpute with additional
latent variables to improve stability, and employs a matrix
decomposition-based knockoff construction that enables efficient inference
in high-dimensional settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HJLee196/scKnockoff}
  \item Report bugs at \url{https://github.com/HJLee196/scKnockoff/issues}
}

}
\author{
\strong{Maintainer}: Hyunjae Lee \email{hyl58@pitt.edu}

Authors:
\itemize{
  \item Lixia Yi \email{LIY70@pitt.edu}
  \item Linxi Liu \email{linxi_liu@pitt.edu}
}

}
